change the file architecture

+++ Puppet/src/Puppet/Core/Timer.h
@@ -15,14 +15,14 @@ namespace Puppet {
 		float Timer::Reset()
 		{
 			std::chrono::high_resolution_clock::time_point end = std::chrono::high_resolution_clock::now();
-			float Mill = std::chrono::duration_cast<std::chrono::milliseconds>(end - m_Start).count();
+			float Mill = (float)std::chrono::duration_cast<std::chrono::milliseconds>(end - m_Start).count();
 			m_Start = end;
 			return Mill*0.001f;
 		}
 		float Timer::Elapsed()
 		{
 			std::chrono::high_resolution_clock::time_point end = std::chrono::high_resolution_clock::now();
-			return std::chrono::duration_cast<std::chrono::milliseconds>(end- m_Start).count();
+			return (float)std::chrono::duration_cast<std::chrono::milliseconds>(end- m_Start).count();
 		}
 		float Timer::ElapsedSeconds()
 		{
